Move postgres_schema -> rail_pm_schema
We're going to have two RDS schemas to maintain, one for the rail
performance manager and one for the metadata. Move the rail performance
manager into its own file thats better named.

Remove the snapshot postgres feature since we don't use it.
