+++ python_src/src/lamp_py/performance_manager/gtfs_utils.py
@@ -7,7 +7,7 @@ import pytz
 import sqlalchemy as sa
 
 from lamp_py.postgres.postgres_utils import DatabaseManager
-from lamp_py.postgres.postgres_schema import (
+from lamp_py.postgres.rail_performance_manager_schema import (
     StaticFeedInfo,
     StaticRoutes,
     StaticStops,
