+++ python_src/src/lamp_py/performance_manager/l0_gtfs_static_mod.py
@@ -2,7 +2,7 @@ import sqlalchemy as sa
 
 
 from lamp_py.postgres.postgres_utils import DatabaseManager
-from lamp_py.postgres.postgres_schema import (
+from lamp_py.postgres.rail_performance_manager_schema import (
     StaticStopTimes,
     StaticStops,
     StaticTrips,
