+++ python_src/src/lamp_py/performance_manager/l1_cte_statements.py
@@ -1,6 +1,6 @@
 import sqlalchemy as sa
 from sqlalchemy.sql.functions import rank
-from lamp_py.postgres.postgres_schema import (
+from lamp_py.postgres.rail_performance_manager_schema import (
     ServiceIdDates,
     StaticStops,
     StaticStopTimes,
