+++ python_src/src/lamp_py/performance_manager/l1_rt_metrics.py
@@ -1,7 +1,7 @@
 import sqlalchemy as sa
 
 from lamp_py.postgres.postgres_utils import DatabaseManager
-from lamp_py.postgres.postgres_schema import (
+from lamp_py.postgres.rail_performance_manager_schema import (
     VehicleEvents,
     TempEventCompare,
 )
