+++ python_src/src/lamp_py/performance_manager/l1_rt_trips.py
@@ -5,7 +5,7 @@ from sqlalchemy.sql.functions import count
 
 
 from lamp_py.postgres.postgres_utils import DatabaseManager
-from lamp_py.postgres.postgres_schema import (
+from lamp_py.postgres.rail_performance_manager_schema import (
     VehicleEvents,
     VehicleTrips,
     StaticTrips,
